Array class constructor and first part of bucket expansion function.
Class constructor that produces the initial array filled with as many objects waiting to be buckets as tell it to.

Also a bucket expansion function that doubles the current array length and generates a new one of that length.

Still need to have it copy the contents of the original array and buckets but that comes after actually put stuff in them.