Partially rewrote most of what already did.
After more troubleshooting and research ended up having to tweak, change, or remove most of what already did.

I think the constructor, set(), and expand() are now working correctly. Only thing unsure about is that when I added enough entries for the list to expand to 28 the highest index was only 13. So the hash() function might need tweaking.

Still need to add duplicate detection/update. After that hopefully the rest will be standard LinkedList manipulation I am already familiar with just with multiple navigated to via hashing for index of which.