Issue with expand resolved.
I think expand() is fully fixed now, the three biggest changes are changing the maxbuckets  from *= to just *,  saving the contents of the current table into a holding array, then emptying the bucketlist before repopulating it with holding.

Next is add duplicate key checking/updating and will be ready to start on the hashtable manipulation functions.